Cast another warning away. The warning was reported by Thomas Zimmermann in https://bugzilla.mozilla.org/show_bug.cgi?id=970271  --This line, and those below, will be ignored--
M    usrsctplib/user_mbuf.h

+++ trunk/KERN/usrsctp/usrsctplib/user_mbuf.h
@@ -165,7 +165,7 @@ struct mbstat {
  */
 #define	MLEN		((int)(MSIZE - sizeof(struct m_hdr)))	/* normal data len */
 #define	MHLEN		((int)(MLEN - sizeof(struct pkthdr)))	/* data len w/pkthdr */
-#define	MINCLSIZE	(MHLEN + 1)	/* smallest amount to put in cluster */
+#define	MINCLSIZE	((int)(MHLEN + 1))	/* smallest amount to put in cluster */
 #define	M_MAXCOMPRESS	(MHLEN / 2)	/* max amount to copy for compression */
